made a few changes that temperarly fixed the issue so now the tasks can be moved

+++ lib/src/main/java/edu/ucsd/cse110/habitizer/lib/domain/Routine.java
@@ -98,8 +98,8 @@ public class Routine implements Serializable {
         );
     }
 
-    public void swapTasks(long id1, long id2){
-        dataSource.swapSortOrders(Math.toIntExact(id1), Math.toIntExact(id2)); //added to communicate to task data source
+    public void swapTasks(int id1, int id2){
+        dataSource.swapSortOrders(id1, id2); //added to communicate to task data source
     }
 
     public int getGoalTime(){
